the big redo - pt1
basically, over the past couple of weeks ive been reworking the whole
backend and some frontend of this project. in the backend i introduced
entities, nhibernate,sessions and a few unittests that i wrote while
reworking code (they just test some basic stuff at this point) and in
frontend i reworked it to not be stupid, complicated and retarded. dafuq
was i thinking before...

next up is getting WPF to work and remove the last bits of old code,
including from the JS

i still have a lot to do, but its going in the correct direction